Move run() method POD to main docs section
`podchecker` noticed that the `=head2` was missing a preceding `=head1`;
upon closer inspection it seemed that the POD for the `run()` method
should have been part of the main documentation section which is at the
end of the file.  Moving the docs for `run()` to the main section fixes
the podchecker warning and puts the docs for this file all in the one
location.